requirements.txt
boto3==1.24.46
botocore==1.27.46
greenlet==1.1.2
jmespath==1.0.1
numpy==1.23.1
pandas==1.4.3
PyMySQL==1.0.2
python-dateutil==2.8.2
pytz==2022.1
s3transfer==0.6.0
six==1.16.0
SQLAlchemy==1.4.39
urllib3==1.26.11
watchtower==3.0.0
